move RenamingAttribute to manager
