+++ src/es6/canvas_widget/Manager.js
@@ -49,6 +49,11 @@ import LogicalOperator from "./viewpoint/LogicalOperator";
 import RenamingListAttribute from "./viewpoint/RenamingListAttribute";
 import ViewTypesUtil from "./viewpoint/ViewTypesUtil";
 
+const renamingAttrHTML = await loadHTML(
+  "../../templates/attribute_widget/renaming_attribute.html",
+  import.meta.url
+);
+
 const singleSelectionAttributeHtml = await loadHTML(
   "../../templates/canvas_widget/single_selection_attribute.html",
   import.meta.url
@@ -8471,6 +8476,158 @@ export class KeySelectionValueAttribute extends AbstractAttribute {
     _$node.find(".value").append(_value.get$node());
   }
 }
+
+/**
+ * RenamingAttribute
+ * @class attribute_widget.ConditionPredicateAttribute
+ * @memberof attribute_widget
+ * @extends attribute_widget.AbstractAttribute
+ * @param {string} id Entity id
+ * @param {string} name Name of attribute
+ * @param {AbstractEntity} subjectEntity Entity the attribute is assigned to
+ * @param {Object} options Selection options
+ * @constructor
+ */
+export class RenamingAttribute extends AbstractAttribute {
+  static TYPE = "RenamingAttribute";
+  constructor(id, name, subjectEntity, options) {
+    super(id, name, subjectEntity);
+
+    //noinspection UnnecessaryLocalVariableJS
+    /**
+     * Selection options
+     * @type {Object}
+     * @private
+     */
+    var _options = options;
+
+    /**
+     * Value object of key
+     * @type {attribute_widget.Value}
+     * @private
+     */
+    var _key = new Value(
+      id + "[val]",
+      "Attribute Name",
+      this,
+      this.getRootSubjectEntity()
+    );
+
+    /***
+     * Value object of ref
+     * @type {attribute_widget.Value}
+     * @private
+     */
+    var _ref = new Value(
+      id + "[ref]",
+      "Attribute Reference",
+      this,
+      this.getRootSubjectEntity()
+    );
+
+    /***
+     * Value object of vis
+     * @type {attribute_widget.Value}
+     * @private
+     */
+    var _vis = new SelectionValue(
+      id + "[vis]",
+      "Attribute Visibility",
+      this,
+      this.getRootSubjectEntity(),
+      _options
+    );
+
+    /**
+     * jQuery object of the DOM node representing the attribute
+     * @type {jQuery}
+     * @private
+     */
+    var _$node = $(_.template(renamingAttrHTML)());
+
+    //noinspection JSUnusedGlobalSymbols
+    /**
+     * Set Value object of key
+     * @param {attribute_widget.Value} key
+     */
+    this.setKey = function (key) {
+      _key = key;
+    };
+
+    /**
+     * Get Value object of key
+     * @returns {attribute_widget.Value}
+     */
+    this.getKey = function () {
+      return _key;
+    };
+
+    /**
+     * Get Value object of value
+     * @returns {attribute_widget.Value}
+     */
+    this.getRef = function () {
+      return _ref;
+    };
+
+    /**
+     * Get Visibility object of value
+     * @returns {attribute_widget.Value}
+     */
+    this.getVis = function () {
+      return _vis;
+    };
+
+    //noinspection JSUnusedGlobalSymbols
+    /**
+     * Set Value object of value
+     * @param {attribute_widget.Value} value
+     */
+    this.setVis = function (value) {
+      _vis = value;
+    };
+
+    /**
+     * Get jQuery object of the DOM node representing the attribute
+     * @returns {jQuery}
+     */
+    this.get$node = function () {
+      return _$node;
+    };
+
+    /**
+     * Set value of key and value by their JSON representation
+     * @param json
+     */
+    this.setValueFromJSON = function (json) {
+      _key.setValueFromJSON(json.val);
+      _ref.setValueFromJSON(json.ref);
+      _vis.setValueFromJSON(json.vis || { value: "" });
+    };
+
+    /**
+     * Get JSON representation of the attribute
+     * @returns {Object}
+     */
+    this.toJSON = function () {
+      var json = AbstractAttribute.prototype.toJSON.call(this);
+      json.val = _key.toJSON();
+      json.ref = _ref.toJSON();
+      json.vis = _vis.toJSON();
+      return json;
+    };
+    _$node.find(".val").append(_key.get$node());
+    _$node.find(".ref").append(_ref.get$node()).hide();
+    _$node.find(".vis").append(_vis.get$node());
+
+    this.registerYMap = function () {
+      _key.registerYType();
+      _ref.registerYType();
+      _vis.registerYType();
+    };
+  }
+}
+
 /**
  * SingleColorValueAttribute
  * @class canvas_widget.SingleColorValueAttribute
